rename to speedbinbreader
