+++ lib-multisrc/speedbinbreader/build.gradle.kts
