+++ src/ja/ohtawebcomic/src/eu/kanade/tachiyomi/extension/ja/ohtawebcomic/OhtaWebComic.kt
@@ -1,6 +1,6 @@
 package eu.kanade.tachiyomi.extension.ja.ohtawebcomic
 
-import eu.kanade.tachiyomi.multisrc.speedbinb.SpeedBinbReader
+import eu.kanade.tachiyomi.multisrc.speedbinbreader.SpeedBinbReader
 import eu.kanade.tachiyomi.network.GET
 import eu.kanade.tachiyomi.network.asObservableSuccess
 import eu.kanade.tachiyomi.source.model.FilterList
